[#12] extract product list from home view
