+++ Expired/Views/ArchivedProductsListView.swift
@@ -1,0 +1,20 @@
+//
+//  ArchivedProductsListView.swift
+//  Expired
+//
+//  Created by Sandeep Singh on 2023-02-13.
+//
+
+import SwiftUI
+
+struct ArchivedProductsListView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct ArchivedProductsListView_Previews: PreviewProvider {
+    static var previews: some View {
+        ArchivedProductsListView()
+    }
+}
